requirements.txt
streamlit
google-genai
chromadb
langchain-text-splitters
sentence-transformers
torch
